% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mrna_lncrna.R
\name{cor_mrna_lncrna}
\alias{cor_mrna_lncrna}
\title{Calculate the correlation between the two expression matrix.}
\usage{
cor_mrna_lncrna(
  filter_log2_mrna,
  filter_log2_lncrna,
  cor_method = "pearson",
  padj_method = "BH",
  digits = 2,
  check_mat = TRUE,
  pstar = FALSE
)
}
\arguments{
\item{filter_log2_mrna}{Expression matrix of mRNA transformed by log2.}

\item{filter_log2_lncrna}{Expression matrix of lncRNA transformed by log2.}

\item{cor_method}{The method of correlation. See \link[stats]{cor}.}

\item{padj_method}{Correction method. See \link[stats]{p.adjust}}

\item{digits}{See \link[base]{round}.}

\item{check_mat}{Logical. Check the expression matrix or not.}

\item{pstar}{Logical. paste stars}
}
\value{
A list
}
\description{
cor_mrna_lncrna(): Calculate the correlation between the two expression matrix.
}
