% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enricherGO.R
\name{enricherGO}
\alias{enricherGO}
\title{enricherGO}
\usage{
enricherGO(
  gene,
  go2ont = NULL,
  ont = "ALL",
  pvalueCutoff,
  qvalueCutoff = 0.2,
  pAdjustMethod = "BH",
  universe,
  minGSSize = 10,
  maxGSSize = 500,
  TERM2GENE,
  TERM2NAME = NA
)
}
\arguments{
\item{gene}{See \link[clusterProfiler]{enricher}.}

\item{go2ont}{a data frame with "go_id’ and "Ontology". If \code{NULL}, it will be downloaded every time. It is recommended to save the \code{go2ont} with the \code{get_GO2Ontology_table()} function first.}

\item{ont}{GO Ontology: "BP", "MF", "CC", "ALL".}

\item{pvalueCutoff}{See \link[clusterProfiler]{enricher}.}

\item{qvalueCutoff}{See \link[clusterProfiler]{enricher}.}

\item{pAdjustMethod}{See \link[clusterProfiler]{enricher}.}

\item{universe}{See \link[clusterProfiler]{enricher}.}

\item{minGSSize}{See \link[clusterProfiler]{enricher}.}

\item{maxGSSize}{See \link[clusterProfiler]{enricher}.}

\item{TERM2GENE}{A data frame with GO IDs and gene IDs. See \link[clusterProfiler]{enricher}.}

\item{TERM2NAME}{A data frame with GO IDs and GO descriptions. See \link[clusterProfiler]{enricher}.}
}
\value{
A object of enrichResult
}
\description{
A more user-friendly enrichment analyzer based on enricher function.
}
\details{
The brother of enricher and enrichGO function.
}
