% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm2tpm.R
\name{fpkm2tpm}
\alias{fpkm2tpm}
\title{Convert FPKM to TPM.}
\usage{
fpkm2tpm(FPKM)
}
\arguments{
\item{FPKM}{A FPKM matrix or data frame of RNA-seq expression.}
}
\value{
A data frame.
}
\description{
Convert FPKM matrix from RNA-seq to TPM matrix.
}
