% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblFasta.R
\name{get_LongestEnsemblFasta}
\alias{get_LongestEnsemblFasta}
\title{get_LongestEnsemblFasta}
\usage{
get_LongestEnsemblFasta(
  fasta,
  source = "Ensembl",
  longest_fa = NULL,
  seq_id = "gene_id",
  width = 60,
  longest_info = NULL
)
}
\arguments{
\item{fasta}{A fasta file of cDNA/pep sequence from Ensembl. Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed.}

\item{source}{A reminder for you to use the Ensembl sequence.}

\item{longest_fa}{Write the longest fasta file. Files ending in .gz, .bz2, .xz, or .zip will be automatically compressed.
if \code{NULL}, it would be ignored.}

\item{seq_id}{Type of fasta IDs (default \code{gene_id}). \code{gene_id}, \code{translation_id}, \code{transcript_id}.}

\item{width}{An integer, (default \code{60}) number of bases of each line in the file.}

\item{longest_info}{Write the information of the longest fasta file. Files ending in .gz, .bz2, .xz, or .zip will be automatically compressed.
if \code{NULL}, it would be ignored.}
}
\value{
A list with two elements containing longest fasta data frame and its all information, \code{fa_longest} and \code{info_longest}.
}
\description{
Get longest fasta from Ensembl cDNA/pep sequence.
}
\details{
Get longest fasta and information from Ensembl cDNA/pep sequence file.
}
