% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst_hrp.R
\docType{data}
\name{pst_hrp}
\alias{pst_hrp}
\title{Sequence data of Pseudomonas syringae pv. tomato hrp gene.}
\format{
A data frame of sequence of Silo COI with 72 rows and 3 variables:
\describe{
\item{id}{The id of sequence}
\item{sequence}{The base sequence of hrp gene}
\item{description}{The description of sequences}
}
}
\usage{
pst_hrp
}
\description{
A dataset containing the ids, sequences and descriptions of Pseudomonas syringae pv. tomato hrp gene
}
\keyword{datasets}
