% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblFasta.R
\name{read_EnsemblFasta}
\alias{read_EnsemblFasta}
\title{read_EnsemblFasta}
\usage{
read_EnsemblFasta(fasta)
}
\arguments{
\item{fasta}{A fasta file from Ensembl, such as cDNA, pep, ncRNA.}
}
\value{
A data frame.
}
\description{
Read fasta file from Ensembl and get its information.
}
\details{
Read fasta file from Ensembl and get its information.
}
