% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_plot_DendroAndColors.R
\name{wgcna_plot_DendroAndColors}
\alias{wgcna_plot_DendroAndColors}
\title{Plot dendrogram and mudule colors.}
\usage{
wgcna_plot_DendroAndColors(
  x,
  abline_height = NULL,
  abline_color = "red",
  n_guide_lines = 50,
  gudide_linetype = "dotted",
  line_interval = 0,
  line_lowest_height = "auto",
  offset_tree_heatmap = 0,
  heamap_height = "auto",
  group_labels = "Mudule colors",
  group_labels_position = "left"
)
}
\arguments{
\item{x}{Output of \code{WGCNA::blockwiseModules()}.}

\item{abline_height}{Abline height. If it is \code{NULL}, abline would not be added.}

\item{abline_color}{The color of abline.}

\item{n_guide_lines}{Number of guide lines. If it is \code{NULL}, guide lines would not be added.}

\item{gudide_linetype}{Specifiy the type of line segment. Learn more about setting these aesthetics in vignette("ggplot2-specs").}

\item{line_interval}{The distance between tips and guide lines.}

\item{line_lowest_height}{Specifiy the lowest height of all guide lines. "auto" or a numberic value.}

\item{offset_tree_heatmap}{The distance between tree object and heatmap.}

\item{heamap_height}{Heatmap height. "auto" or a numberic value.}

\item{group_labels}{Label for the heatmap.}

\item{group_labels_position}{Position of \code{group_labels}. "left" or "right".}
}
\value{
A ggtree object.
}
\description{
Plot dendrogram and mudule colors using ggtree. This function is similar to WGCNA::plotDendroAndColors() and
it origined from \code{ggtreeDendro::plot_wgcna()}.
}
