% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bio_fasta.R
\name{write_bio_fasta}
\alias{write_bio_fasta}
\title{Write a data frame to a file in fasta format.}
\usage{
write_bio_fasta(
  data,
  file,
  width = 60,
  append = FALSE,
  num_threads = readr::readr_threads()
)
}
\arguments{
\item{data}{A data frame with 2 or 3 columns including \code{id}, \code{sequence} and \code{description},separately.}

\item{file}{A path to a file of fasta sequence.}

\item{width}{An integer, (default \code{60}) number of bases of each line in the file.}

\item{append}{see \link[readr]{write_lines}. If \code{FALSE}, will overwrite existing file. If \code{TRUE}, will append to existing file. In both cases, if the file does not exist a new file is created.}

\item{num_threads}{see \link[readr]{write_lines}. The number of processing threads to use for initial parsing and lazy reading of data. If your data contains newlines within fields the parser should automatically detect this and fall back to using one thread only. However if you know your file has newlines within quoted fields it is safest to set \code{num_threads = 1} explicitly.}
}
\value{
A file in fasta format.
}
\description{
\code{write_bio_fasta()}: write a data frame to a file in fasta format.
}
\examples{
\dontrun{
data(pst_hrp)
write_bio_fasta(data = pst_hrp, file = "pst_hrp.fasta")
}
}
