% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cor_p.R
\name{write_cor_p}
\alias{write_cor_p}
\title{Write the result of \code{cor_mrna_lncrna()} to disk.}
\usage{
write_cor_p(
  cor_p,
  r_cutoff = 0.9,
  padj_cutoff = 0.05,
  output_path = "./",
  prefix = "my",
  num_threads = 5,
  gzip = TRUE,
  get_WiderData = FALSE
)
}
\arguments{
\item{cor_p}{A list from \link[BioUncle]{cor_mrna_lncrna}.}

\item{r_cutoff}{r cutoff.}

\item{padj_cutoff}{Padj cutoff.}

\item{output_path}{The path of results files.}

\item{prefix}{The prefix of results files.}

\item{num_threads}{Number of threads, see\link[vroom]{vroom_write}.}

\item{gzip}{Logical, compress the output or not.}

\item{get_WiderData}{Logical, write wide data frame or not.}
}
\value{
Result files.
}
\description{
write_cor_p(): write the result of \code{cor_mrna_lncrna()} to disk.
}
