% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblFasta.R
\name{write_LongestEnsemblFasta}
\alias{write_LongestEnsemblFasta}
\title{write_LongestEnsemblFasta}
\usage{
write_LongestEnsemblFasta(list_longest, longest_fa, width = 60, longest_info)
}
\arguments{
\item{list_longest}{The result list of get_LongestEnsemblFasta function.}

\item{longest_fa}{Write the longest fasta file. Files ending in .gz, .bz2, .xz, or .zip will be automatically compressed.
if \code{NULL}, it would be ignored.}

\item{width}{An integer, (default \code{60}) number of bases of each line in the file.}

\item{longest_info}{Write the information of the longest fasta file. Files ending in .gz, .bz2, .xz, or .zip will be automatically compressed.
if \code{NULL}, it would be ignored.}
}
\description{
Write the fasta file of longest sequence and its information.
}
