% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one2mul.R
\name{get_one2mul}
\alias{get_one2mul}
\title{Convert to single cell according to the same content.}
\usage{
get_one2mul(df, sep = "/")
}
\arguments{
\item{df}{A data frame with two columns.}

\item{sep}{A separator character.}
}
\value{
A data frame with two character vectors.
}
\description{
A function opposite to \link[tidyr]{separate_rows}.
}
\examples{
\dontrun{
df <- iris \%>\% select(5,3)
str(df)
new_df <- get_one2mul(df)

str(new_df)
}
}
